refactor(tui): separate input stack contracts
